% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineSCEs2MAE.R
\name{combineSCEs2MAE}
\alias{combineSCEs2MAE}
\title{Combine SingleCellExperiments into a MultiAssayExperiment}
\usage{
combineSCEs2MAE(
  atac.sce = NULL,
  rna.sce = NULL,
  sample.annotation = NULL,
  use.alt.exp = FALSE,
  main.exp.name = "TileMatrix500"
)
}
\arguments{
\item{atac.sce}{A \linkS4class{SingleCellExperiment} containing ATAC data.}

\item{rna.sce}{A \linkS4class{SingleCellExperiment} containing RNA data.}

\item{sample.annotation}{A data.frame containing sample annotation.}

\item{use.alt.exp}{Logical for selecting the MultiAssayExperiment structure.}

\item{main.exp.name}{String containing the name of the experiment that will
be the main experiment when use.alt.exp is TRUE.}
}
\value{
A \linkS4class{MultiAssayExperiment}
}
\description{
Combine ATAC and RNA SingleCellExperiments into a MultiAssayExperiment.
}
\author{
Natalie Fox
}
