% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createATACSCE.R
\name{createATACSCE}
\alias{createATACSCE}
\title{Create a SingleCellExperiment from fragment files}
\usage{
createATACSCE(
  fragment.files,
  output.dir = tempdir(),
  tile.size = 500,
  seq.lengths = NULL,
  barcodes.list = NULL,
  BPPARAM = bpparam()
)
}
\arguments{
\item{fragment.files}{Vector of strings specifying fragment files. Vector
names need to be sample names.}

\item{output.dir}{String containing the directory where files should be
output.}

\item{tile.size}{Integer scalar specifying the size of the tiles in base
pairs.}

\item{seq.lengths}{Named integer vector containing the lengths of the
reference sequences used for alignment.}

\item{barcodes.list}{A List with samples as names and the values a vector of
barcodes for that sample.}

\item{BPPARAM}{A \linkS4class{BiocParallelParam} object indicating how matrix
creation should be parallelized.}
}
\value{
A \linkS4class{SingleCellExperiment} containing the tile matrix.
}
\description{
Create a SingleCellExperiment from fragment files, storing the tile matrix.
}
\examples{
temp1 <- tempfile(fileext=".fragments.gz")
mockFragmentFile(temp1, c(chrA=1000, chrB=200000, chrC=200),
                 num.fragments=100, cell.names=LETTERS)
temp2 <- tempfile(fileext=".fragments.gz")
mockFragmentFile(temp2, c(chrA=1000, chrB=200000, chrC=200),
                 num.fragments=100, cell.names=LETTERS)
fragment_files <- c(sample1=temp1, sample2=temp2)
tile_sce <- createATACSCE(fragment.files = fragment_files, tile.size=500, 
                          seq.lengths = c(chrA=1001, chrB=200001, chrC=201))

}
\author{
Natalie Fox
}
