% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMAE.R
\name{createMAEFromCellranger}
\alias{createMAEFromCellranger}
\title{Import a single cell multiome MultiAssayExperiment}
\usage{
createMAEFromCellranger(
  cellranger.dirs,
  sample.names,
  output.dir = tempdir(),
  tile.size = 500,
  seq.lengths = NULL,
  regions = NULL,
  atac.name = "TileMatrix500",
  rna.name = "GeneExpressionMatrix",
  fragment.file = "atac_fragments.tsv.gz",
  h5.file = "filtered_feature_bc_matrix.h5",
  barcodes.list = NULL,
  consistent.barcodes = TRUE,
  sample.annotation = NULL,
  filter.features.without.intervals = TRUE,
  BPPARAM = bpparam()
)
}
\arguments{
\item{cellranger.dirs}{Character vector specifying Cell Ranger multiome analysis directories}

\item{sample.names}{Character vector specifying sample names corresponding
to fragment.files}

\item{output.dir}{String containing the directory where files should be
output}

\item{tile.size}{Integer scalar specifying the size of the tiles in base
pairs}

\item{seq.lengths}{Named integer vector containing the lengths of the
reference sequences used for alignment. Vector names should correspond to
the names of the sequences, in the same order of occurrence as in the
fragment file. If \code{NULL}, this is obtained from the reference genome
used by Cellranger (itself located by scanning the header of the fragment
file).}

\item{regions}{Genomic Ranges specifying gene coordinates for creating the
gene score matrix.}

\item{atac.name}{String indicating the name of \code{atac.sce}}

\item{rna.name}{String indicating the name of \code{rna.sce}}

\item{fragment.file}{Character string indicating the name of the fragment file}

\item{h5.file}{Character string indicating the name of feature matrix h5 file}

\item{barcodes.list}{A named list with samples as names and each list element
as a character vector of barcodes. If \code{NULL}, barcodes are extracted
from h5.file}

\item{consistent.barcodes}{Logical scalar indicating whether only shared barcodes should be
kept}

\item{sample.annotation}{A data.frame containing sample annotation.}

\item{filter.features.without.intervals}{Logical scalar indicating whether to remove
.   features from the h5.files that do not have interval specified. Often these are
   mitochondria genes.}

\item{BPPARAM}{A \link[BiocParallel]{BiocParallelParam} object indicating how matrix
creation should be parallelized.}
}
\value{
A \link[MultiAssayExperiment]{MultiAssayExperiment}.
}
\description{
Import results from single cell multiome Cell Ranger results directories
into a MultiAssayExperiment object.
}
\examples{
# create mock cell range output files
tmp_dir <- tempdir()
temp_rna <- file.path(tmp_dir,"filtered_feature_bc_matrix.h5")
mockCellRangerH5(temp_rna, cell.names = LETTERS)
temp_atac <- file.path(tmp_dir,"atac_fragments.tsv.gz")
mockFragmentFile(temp_atac, 
                 c(chrA=1000, chrB=200000, chrC=200),
                 num.fragments=100, 
                 cell.names=LETTERS)
seq.lengths <-  c(chrA=1000, chrB=200000, chrC=200)
# create combined MAE
mae <- createMAEFromCellranger(cellranger.dirs = tmp_dir,
                               sample.names = "mock",
                               seq.lengths = seq.lengths,
                               consistent.barcodes = FALSE)
file.remove(list.files(tmp_dir, pattern=".h5", full.names = TRUE))
}
\author{
Natalie Fox, Xiaosai Yao
}
