% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createATACSCE.R
\name{createRegionSCE}
\alias{createRegionSCE}
\title{Create a SingleCellExperiment for Gene Scores from fragment files}
\usage{
createRegionSCE(
  fragment.files,
  sample.names,
  output.dir = tempdir(),
  regions,
  barcodes.list = NULL,
  matrix.name = "GeneAccessibilityMatrix",
  BPPARAM = bpparam()
)
}
\arguments{
\item{fragment.files}{Character vector specifying fragment files}

\item{sample.names}{Character vector specifying sample names corresponding
to fragment.files}

\item{output.dir}{String containing the directory where files should be
output}

\item{regions}{Genomic Ranges specifying gene coordinates for creating the
gene score matrix.}

\item{barcodes.list}{A named list with samples as names and each list element
as a character vector of barcodes. If \code{NULL}, all barcodes are extracted}

\item{matrix.name}{Character string indicating the name of the matrix}

\item{BPPARAM}{A \link[BiocParallel]{BiocParallelParam} object indicating how matrix
creation should be parallelized.}
}
\value{
A \link[SingleCellExperiment]{SingleCellExperiment} containing the gene score matrix.
}
\description{
Create a SingleCellExperiment from fragment files, storing the gene score matrix.
}
\examples{
temp1 <- tempfile(fileext=".fragments.gz")
mockFragmentFile(temp1, c(chrA=1000, chrB=200000, chrC=200),
                 num.fragments=100, cell.names=LETTERS)
temp2 <- tempfile(fileext=".fragments.gz")
mockFragmentFile(temp2, c(chrA=1000, chrB=200000, chrC=200),
                 num.fragments=100, cell.names=LETTERS)
fragment_files <- c(sample1=temp1, sample2=temp2)
region_sce <- createRegionSCE(fragment.files=c(temp1,temp2),
                              sample.names=c("sample1", "sample2"),                                
                              region=GenomicRanges::GRanges(c("chrA:500-1000", "chrB:1000-2000")))
file.remove(list.files(tempdir(), pattern=".h5", full.names = TRUE))
}
\author{
Natalie Fox, Jayaram Kancherla, Xiaosai Yao
}
