% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockInputFiles.R
\name{mockCellRangerH5}
\alias{mockCellRangerH5}
\title{Create a mock Cell Ranger h5 file}
\usage{
mockCellRangerH5(filepath, n.genes = 100, n.cells = 50, cell.names = NULL)
}
\arguments{
\item{filepath}{String containing a path to the output file.}

\item{n.genes}{Integer indicating the number of genes}

\item{n.cells}{Integer indicating the number of cells}

\item{cell.names}{Character vector containing the cell names. The length of
this vector is used as the total number of cells.}
}
\value{
A h5 file mimicking a typical Cell Ranger output is created at \code{filepath}.
}
\description{
Mock up a Cell Ranger h5 file for examples and testing.
}
\examples{
temp <- tempfile(fileext = ".h5")
mockCellRangerH5(temp, cell.names = LETTERS)
h5ls(temp)

}
\author{
Jayaram Kancherla
}
