% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMAE.R
\name{createArbalistMAE}
\alias{createArbalistMAE}
\title{Combine SingleCellExperiments into a MultiAssayExperiment}
\usage{
createArbalistMAE(
  atac.sce,
  rna.sce,
  sample.annotation = NULL,
  rna.name = "GeneExpressionMatrix",
  atac.name = "TileMatrix500"
)
}
\arguments{
\item{atac.sce}{A \link[SingleCellExperiment]{SingleCellExperiment} object
containing ATAC data}

\item{rna.sce}{A \link[SingleCellExperiment]{SingleCellExperiment} object
containing RNA data}

\item{sample.annotation}{A data.frame containing sample annotation.}

\item{rna.name}{String indicating the name of \code{rna.sce}}

\item{atac.name}{String indicating the name of \code{atac.sce}}
}
\value{
A \link[MultiAssayExperiment]{MultiAssayExperiment}.
}
\description{
Combine ATAC and RNA SingleCellExperiments into a MultiAssayExperiment.
}
\examples{
# create rna sce
temp_rna <- tempfile(fileext = ".h5")
mockCellRangerH5(temp_rna, cell.names = LETTERS)
rna_sce <- createRNASCE(temp_rna, sample.names="sample1")

# create atac sce
temp_atac <- tempfile(fileext=".fragments.gz")
mockFragmentFile(temp_atac, c(chrA=1000, chrB=200000, chrC=200),
                 num.fragments=100, cell.names=LETTERS)
tile_sce <- createTileSCE(fragment.files=temp_atac,
                          sample.names="sample1",
                          tile.size=500,
                          seq.lengths = c(chrA=1000, chrB=200000, chrC=200))
# create combined MAE
mae <- createArbalistMAE(tile_sce, rna_sce)
file.remove(list.files(tempdir(), pattern=".h5", full.names = TRUE))
}
\author{
Natalie Fox, Jayaram Kancherla
}
