% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMAE.R
\name{createMAEFromCellranger}
\alias{createMAEFromCellranger}
\title{Import a single cell ATAC-seq MultiAssayExperiment}
\usage{
createMAEFromCellranger(
  cellranger.dirs,
  sample.names,
  output.dir = tempdir(),
  tile.size = 500,
  seq.lengths = NULL,
  regions = NULL,
  atac.name = "TileMatrix500",
  rna.name = "GeneExpressionMatrix",
  fragment.file = "atac_fragments.tsv.gz",
  h5.file = "filtered_feature_bc_matrix.h5",
  barcode.file = "per_barcode_metrics.csv",
  summary.file = "summary.csv",
  filter.features.without.intervals = TRUE,
  BPPARAM = bpparam()
)
}
\arguments{
\item{cellranger.dirs}{Named character vector specifying a Cell Ranger
scATAC-seq or multiome results directory. Vector names need to be sample
names.}

\item{sample.names}{A character vector specifying sample names. These will be 
added in front of the barcodes}

\item{output.dir}{String containing the directory where files should be
output.}

\item{tile.size}{Integer scalar specifying the size of the tiles in base
pairs.}

\item{seq.lengths}{Named integer vector containing the lengths of the
reference sequences used for alignment.}

\item{regions}{Genomic Ranges specifying gene coordinates for creating the
gene score matrix.}

\item{atac.name}{String indicating the name of \code{atac.sce}}

\item{rna.name}{String indicating the name of \code{rna.sce}}

\item{fragment.file}{Character string indicating the name of the fragment file}

\item{h5.file}{Character string indicating the name of feature matrix h5 file}

\item{barcode.file}{Character string indicating the name of the barcode file}

\item{summary.file}{Character string indicating the name of the summary file}

\item{filter.features.without.intervals}{Logical scalar indicating whether to remove
.   features from the h5.files that do not have interval specified. Often these are
   mitochondria genes.}

\item{BPPARAM}{A \link[BiocParallel]{BiocParallelParam} object indicating how matrix
creation should be parallelized.}
}
\value{
A \link[MultiAssayExperiment]{MultiAssayExperiment}.
}
\description{
Import results from scATAC-seq or multiome Cell Ranger results directories
into a MultiAssayExperiment.
}
\author{
Natalie Fox, Xiaosai Yao
}
