% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createATACSCE.R
\name{createRegionSCE}
\alias{createRegionSCE}
\title{Create a SingleCellExperiment for Gene Scores from fragment files}
\usage{
createRegionSCE(
  fragment.files,
  regions,
  output.dir = tempdir(),
  barcodes.list = NULL,
  BPPARAM = bpparam()
)
}
\arguments{
\item{fragment.files}{Vector of strings specifying fragment files. Vector
names need to be sample names.}

\item{regions}{Genomic Ranges specifying gene coordinates for creating the
gene score matrix.}

\item{output.dir}{String containing the directory where files should be
output.}

\item{barcodes.list}{A List with samples as names and the values a vector of
barcodes for that sample.}

\item{BPPARAM}{A \linkS4class{BiocParallelParam} object indicating how matrix
creation should be parallelized.}
}
\value{
A \linkS4class{SingleCellExperiment} containing the gene score matrix.
}
\description{
Create a SingleCellExperiment from fragment files, storing the gene score matrix.
}
\examples{
temp1 <- tempfile(fileext=".fragments.gz")
mockFragmentFile(temp1, c(chrA=1000, chrB=200000, chrC=200),
                 num.fragments=100, cell.names=LETTERS)
temp2 <- tempfile(fileext=".fragments.gz")
mockFragmentFile(temp2, c(chrA=1000, chrB=200000, chrC=200),
                 num.fragments=100, cell.names=LETTERS)
fragment_files <- c(sample1=temp1, sample2=temp2)
region_sce <- createGeneScoreSCE(fragment.files = fragment_files, 
                          genes.grs = GRanges(c("chrA:500-1000", "chrB:1000-2000")))

}
\author{
Natalie Fox
}
