% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRNASCE.R
\name{createRNASCE}
\alias{createRNASCE}
\title{Import gene expression results from multiome Cell Ranger results}
\usage{
createRNASCE(
  h5.files,
  sample.names = NULL,
  feature.type = "Gene Expression",
  exp.name = "GeneExpressionMatrix",
  filter.features.without.intervals = TRUE
)
}
\arguments{
\item{h5.files}{Character vector specifying path to a
filtered_feature_bc_matrix.h5. Vector must be the same length as
\code{sample.names}.}

\item{sample.names}{Character vector specifying sample names. Vector must be the
same length as \code{h5.files}.}

\item{feature.type}{String specifying the feature type to select from
filtered_feature_bc_matrix.h5.}

\item{exp.name}{String specifying the name of the experiment}

\item{filter.features.without.intervals}{Logical scalar indicating whether to remove
.   features from the h5.files that do not have interval specified. Often these are
   mitochondria genes.}
}
\value{
A \link[SingleCellExperiment]{SingleCellExperiment}
}
\description{
Import RNA results from multiome Cell Ranger results into a
SingleCellExperiment.
}
\examples{
temp <- tempfile(fileext = ".h5")
mockCellRangerH5(temp, cell.names = LETTERS)
rna_sce <- createRNASCE(temp, sample.names="sample1")
file.remove(list.files(tempdir(), pattern=".h5", full.names = TRUE))

}
\author{
Natalie Fox
}
