% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createATACSCE.R
\name{createTileSCE}
\alias{createTileSCE}
\title{Create a SingleCellExperiment from fragment files}
\usage{
createTileSCE(
  fragment.files,
  sample.names,
  output.dir = tempdir(),
  tile.size = 500,
  seq.lengths = NULL,
  barcodes.list = NULL,
  matrix.name = "TileMatrix",
  BPPARAM = bpparam()
)
}
\arguments{
\item{fragment.files}{Character vector specifying fragment files}

\item{sample.names}{Character vector specifying sample names corresponding
to fragment.files}

\item{output.dir}{String containing the directory where files should be
output}

\item{tile.size}{Integer scalar specifying the size of the tiles in base
pairs}

\item{seq.lengths}{Named integer vector containing the lengths of the
reference sequences used for alignment. Vector names should correspond to
the names of the sequences, in the same order of occurrence as in the
fragment file. If \code{NULL}, this is obtained from the reference genome
used by Cellranger (itself located by scanning the header of the fragment
file).}

\item{barcodes.list}{A named list with samples as names and each list element
as a character vector of barcodes. If \code{NULL}, all barcodes are extracted}

\item{matrix.name}{Character string indicating the name of the matrix}

\item{BPPARAM}{A \link[BiocParallel]{BiocParallelParam} object indicating how matrix
creation should be parallelized.}
}
\value{
A \link[SingleCellExperiment]{SingleCellExperiment} containing the tile matrix.
}
\description{
Create a SingleCellExperiment from fragment files, storing the tile matrix.
}
\examples{
temp1 <- tempfile(fileext=".fragments.gz")
mockFragmentFile(temp1, c(chrA=1000, chrB=200000, chrC=200),
                 num.fragments=100, cell.names=LETTERS)
temp2 <- tempfile(fileext=".fragments.gz")
mockFragmentFile(temp2, c(chrA=1000, chrB=200000, chrC=200),
                 num.fragments=100, cell.names=LETTERS)
tile_sce <- createTileSCE(fragment.files=c(temp1,temp2), 
                          sample.names=c("sample1", "sample2"),
                          tile.size=500, 
                          seq.lengths = c(chrA=1000, chrB=200000, chrC=200))
file.remove(list.files(tempdir(), pattern=".h5", full.names = TRUE))
}
\author{
Natalie Fox, Jayaram Kancherla, Xiaosai Yao
}
