% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractGenomeRefFilenameFromFragmentFile.R
\name{extractGenomeRefFilenameFromFragmentFile}
\alias{extractGenomeRefFilenameFromFragmentFile}
\title{Retrieves genome reference file name}
\usage{
extractGenomeRefFilenameFromFragmentFile(fragment.file)
}
\arguments{
\item{fragment.file}{String specifying fragment file name}
}
\value{
Named string vector
}
\description{
Helper function to extract genome reference file name from fragment file header.
}
\examples{
# Mock a fragment file
f <- tempfile(fileext=".tsv.gz")
mockFragmentFile(f, c(chr1=1000), 10, LETTERS[1:5], 
                 comments=c("reference_path=/path/to/ref"))

extractGenomeRefFilenameFromFragmentFile(f)

}
\author{
Natalie Fox
}
