% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filterDuplicateFeatures}
\alias{filterDuplicateFeatures}
\title{Filter duplicate features}
\usage{
filterDuplicateFeatures(
  se,
  mcol.name = "name",
  assay.name = 1,
  summary.stat = sum,
  selection.metric = max
)
}
\arguments{
\item{se}{A \link[SummarizedExperiment]{SummarizedExperiment} object containing
RNA data}

\item{mcol.name}{String specifying the column name in the rowData 
corresponding to feature names}

\item{assay.name}{String specifying the name of the assay}

\item{summary.stat}{Function to summarize each feature (row) of the
experiment. Default is \code{sum}}

\item{selection.metric}{Function to select the row to keep when there are
duplicate rows with the same \code{mcol.name}. Default is \code{max}}
}
\value{
A \link[SummarizedExperiment]{SummarizedExperiment} with duplicate features resolved.
}
\description{
Keeps one feature from each set of duplicated features based on specified
criteria, such as the feature with the highest values. Other duplicate
features are removed.
}
\examples{
library(SummarizedExperiment)

# Create a mock SummarizedExperiment with duplicates
counts <- matrix(1:10, ncol=2)
rdata <- DataFrame(name = c("GeneA", "GeneA", "GeneB", "GeneC", "GeneC"))
se <- SummarizedExperiment(assays=list(counts=counts), rowData=rdata)

# Filter duplicates, keeping the row with the max sum
se_filtered <- filterDuplicateFeatures(se, mcol.name="name", selection.metric=max)

}
\author{
Natalie Fox, Jayaram Kancherla
}
