% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExpListFromFragments.R
\name{getSCEFromH5List}
\alias{getSCEFromH5List}
\title{Create a SingleCellExperiment from a list of delayed matrices}
\usage{
getSCEFromH5List(h5.res.list, grs)
}
\arguments{
\item{h5.res.list}{A list containing delayed matrices with HDF5 backends that will be combined using AmalgamatedArray into a SingleCellExperiment. List item names should be the sample name for the delayed matrix.}

\item{grs}{GRange object to be used for the rowRanges of the resulting SingleCellExperiment}
}
\value{
A SingleCellExperiment
}
\description{
Create a SingleCellExperiment from a list of delayed matrices using AmalgamatedArray.
}
\author{
Natalie Fox
}
